Initial attempt at getting source info into form metadata
Top-level metadatable forms read by the reader will now have correct
:line and :column keys in their metadata, indicating the places at
which they were defined in their defining source files. Descendants of
these top-level forms will have incorrect :line and :column keys,
because of the way in which the reader tracks source information; this
ought to be fixed in the near future.